pyproject.toml
[project]
name = "graphql-hands-on"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.121.1",
    "strawberry-graphql>=0.283.3",
    "uvicorn>=0.38.0",
]
